pyproject.toml
[project]
name = "leaps"
version = "2.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate==1.10.1",
    "biopython==1.85",
    "datasets==4.0.0",
    "easydict==1.13",
    "ipykernel==6.30.1",
    "lightning==2.5.5",
    "nbstripout==0.8.1",
    "numpy==2.3.2",
    "optuna==4.5.0",
    "pandas==2.3.2",
    "peft==0.17.1",
    "pyyaml==6.0.2",
    "safetensors==0.6.2",
    "scikit-learn==1.7.1",
    "torch==2.8.0",
    "torchmetrics==1.8.2",
    "tqdm==4.67.1",
    "transformers==4.56.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]
